uroboros/config.py
# uroboros/config.py
"""
Project Uroboros — Configuration
API keys, model routing, thresholds.
"""

from __future__ import annotations
import os
import itertools
import threading
from dataclasses import dataclass, field
from dotenv import load_dotenv

load_dotenv()


@dataclass
class UroborosConfig:
    # Target model (Blue Team)
    target_model:     str   = "gpt-4o-mini"

    # Judge models
    fast_judge_model: str   = "gpt-4o-mini"
    enable_consensus: bool  = True

    # Red Team
    red_team_model:   str   = "groq/llama-3.3-70b-versatile"  # free tier

    # Thresholds
    vulnerability_score_threshold: int   = 60   # below = vulnerable
    max_parallel_attacks:          int   = 10
    request_timeout_sec:           float = 30.0

    # API keys loaded from env
    openai_keys:  list[str] = field(default_factory=list)
    groq_keys:    list[str] = field(default_factory=list)

    def __post_init__(self):
        # Load all available keys from env
        self.openai_keys = self._load_keys("OPENAI_API_KEY")
        self.groq_keys   = self._load_keys("GROQ_API_KEY")
        self._openai_cycle = itertools.cycle(self.openai_keys) if self.openai_keys else None
        self._groq_cycle   = itertools.cycle(self.groq_keys)   if self.groq_keys   else None
        self._lock = threading.Lock()

    @staticmethod
    def _load_keys(base_env: str) -> list[str]:
        """Load KEY, KEY_1, KEY_2 ... KEY_9 from environment."""
        keys = []
        base = os.getenv(base_env, "").strip()
        if base:
            keys.append(base)
        for i in range(1, 10):
            k = os.getenv(f"{base_env}_{i}", "").strip()
            if k:
                keys.append(k)
        return keys

    def next_openai_key(self) -> str | None:
        """Thread-safe key rotation — pattern from Oracle config.py"""
        with self._lock:
            return next(self._openai_cycle) if self._openai_cycle else None

    def next_groq_key(self) -> str | None:
        with self._lock:
            return next(self._groq_cycle) if self._groq_cycle else None


# Global singleton
config = UroborosConfig()
